requirements.txt
python-dotenv>=1.0.1
groq>=0.9.0
httpx>=0.27.0
PyPDF2>=3.0.1
beautifulsoup4>=4.12.3
mypy-extensions>=1.0.0
types-requests; python_version>='3.8'
types-python-dateutil; python_version>='3.8'
pytest>=8.2.0
pytest-cov>=5.0.0
rich>=13.7.1
uvloop; platform_system != 'Windows'
markdown>=3.6
